+++ test/docfx.Snapshot.Tests/ModuleInitializer.cs
@@ -1,0 +1,20 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.DocAsCode.Tests
+{
+    using System.Runtime.CompilerServices;
+    using VerifyTests;
+
+    public static class ModuleInitializer
+    {
+        [ModuleInitializer]
+        public static void Initialize()
+        {
+            VerifyDiffPlex.Initialize(VerifyTests.DiffPlex.OutputType.Compact);
+
+            VerifyImageMagick.Initialize();
+            VerifyImageMagick.RegisterComparers(0.05);
+        }
+    }
+}
